examples/CH_09/requirements.txt
alembic==1.7.5
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.3.0
dnspython==2.1.0
dynaconf==3.1.7
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.9.1
mako==1.2.2
MarkupSafe==2.0.1
premailer==3.10.0
pycparser==2.21
pyjwt==2.4.0
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
waitress==2.1.2
Werkzeug==2.0.2
WTForms==3.0.0
sib-api-v3-sdk==7.5.0